+++ src/hooks/useUserInfo.tsx
@@ -1,0 +1,4 @@
+import { UserInfoContext } from '@/contexts/UserInfoContext';
+import { useContext } from 'react';
+
+export const useUserInfo = () => useContext(UserInfoContext);
