+++ src/routes/Router.tsx
@@ -1,0 +1,21 @@
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import GiftShop from '@/pages/GiftShop';
+import Login from '@/pages/Login';
+import NotFound from '@/pages/NotFound';
+import { UserInfoProvider } from '@/providers/UserInfoProvider';
+
+const Router = () => {
+  return (
+    <UserInfoProvider>
+      <BrowserRouter>
+        <Routes>
+          <Route path="/" element={<GiftShop />} />
+          <Route path="/login" element={<Login />} />
+          <Route path="*" element={<NotFound />} />
+        </Routes>
+      </BrowserRouter>
+    </UserInfoProvider>
+  );
+};
+
+export default Router;
