+++ src/styles/global.tsx
@@ -6,8 +6,7 @@ const GlobalStyle = () => (
     styles={css`
       @import url('https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard.css');
 
-      html,
-      body {
+      html {
         font-family: 'Pretendard', sans-serif;
         background-color: #f7f8f9;
       }
